refactor: rpc namespace refactor - net
